requirements.txt
pandas
numpy
openai
langchain
pypdf
rank_bm25
tiktoken
faiss-cpu
spacy
sentence-transformers